feat(cactus-icons): Export styled svgs directly, fix svgs with stroke set
