+++ modules/cactus-icons/package.json
@@ -32,6 +32,9 @@
     "transform": {
       "^.+\\.(t|j)sx?$": "babel-jest"
     },
+    "setupFilesAfterEnv": [
+      "<rootDir>/tests/__setup__/setup.ts"
+    ],
     "testRegex": ".*\\.test\\.(tsx?)$",
     "verbose": true,
     "coverageReporters": [
@@ -59,6 +62,7 @@
     "babel-jest": "^24.7.1",
     "fast-glob": "^2.2.6",
     "jest": "^24.7.1",
+    "jest-styled-components": "^6.2.1",
     "prettier": "^1.17.0",
     "react": "^16.8.6",
     "react-dom": "^16.8.6",
