+++ modules/cactus-icons/tests/__setup__/setup.ts
@@ -1,0 +1,1 @@
+import 'jest-styled-components'
